+++ docs/src/components/preview.svelte
@@ -1,87 +1,10 @@
-<script context="module" lang="ts">
-	type BooleanControl = {
-		label: string;
-		defaultValue: boolean;
-		type: "boolean";
-	};
-
-	type SelectControl = {
-		label: string;
-		defaultValue: string;
-		options: string[];
-		type: "select";
-	};
-
-	type NumberControl = {
-		label: string;
-		defaultValue: number;
-		type: "number";
-		min?: number;
-		max?: number;
-	};
-
-	type Control = BooleanControl | SelectControl | NumberControl;
-
-	type NormalizeType<T> = T extends string
-		? T
-		: T extends number
-			? T
-			: T extends boolean
-				? boolean
-				: // eslint-disable-next-line @typescript-eslint/no-explicit-any
-					T extends Record<string, any>
-					? T
-					: never;
-
-	type SchemaExtends = Record<string, Control>;
-
-	type Context<Schema extends SchemaExtends> = {
-		values: {
-			[K in keyof Schema]: NormalizeType<
-				Schema[K] extends SelectControl ? Schema[K]["options"][number] : Schema[K]["defaultValue"]
-			>;
-		};
-		schema: Schema;
-	};
-
-	const ctx = {
-		set<Schema extends SchemaExtends>(ctx: Context<Schema>) {
-			return setContext(CTX_KEY, ctx);
-		},
-		get<Schema extends SchemaExtends>() {
-			return getContext<Context<Schema>>(CTX_KEY) ?? {};
-		},
-	};
-
-	const CTX_KEY = Symbol();
-
-	// A type that marks all readonly values as writable
-	type Writable<T> = {
-		-readonly [P in keyof T]: T[P];
-	};
-
-	export function usePreviewControls<const Schema extends SchemaExtends>(
-		schema: Schema,
-	): Writable<Context<Schema>["values"]> {
-		const values = $state(
-			objectMap(schema, (key, { defaultValue }) => {
-				return [key, defaultValue];
-			}),
-		) as Context<Schema>["values"];
-
-		ctx.set<Schema>({ values, schema });
-
-		return values;
-	}
-</script>
-
 <script lang="ts">
-	import { objectMap } from "@antfu/utils";
-	import { getContext, setContext, type Snippet } from "svelte";
 	import { linear } from "svelte/easing";
 	import { fade, type TransitionConfig } from "svelte/transition";
+	import { previewCtx } from "./preview-ctx.svelte";
+	import type { Snippet } from "svelte";
 
-	const { values, schema } = ctx.get();
+	const { values, schema } = previewCtx.get();
 
 	interface Props {
 		children: Snippet;
@@ -141,7 +64,7 @@
 		{@render children()}
 	</div>
 
-	{#if !open}
+	{#if !open && values}
 		<button
 			class="absolute bottom-4 left-4 z-10 cursor-pointer rounded-lg bg-gray-500 px-2 py-1
 		text-sm text-white transition hover:bg-gray-700 active:bg-gray-800"
@@ -182,7 +105,7 @@
 					{:else if control.type === "select"}
 						<select
 							bind:value={values[key] as string}
-							class="self-stretch rounded-md bg-gray-900 px-1 py-0.5"
+							class="self-stretch rounded-md px-1 py-0.5 dark:bg-gray-900"
 						>
 							{#each control.options as option}
 								<option value={option}>{option}</option>
@@ -194,7 +117,13 @@
 							bind:value={values[key] as number}
 							min={control.min}
 							max={control.max}
-							class="self-stretch rounded-md bg-gray-900 px-1 py-0.5"
+							class="self-stretch rounded-md px-1 py-0.5 dark:bg-gray-900"
+						/>
+					{:else if control.type === "string"}
+						<input
+							type="text"
+							bind:value={values[key] as string}
+							class="self-stretch rounded-md px-1 py-0.5 dark:bg-gray-900"
 						/>
 					{/if}
 				</label>