+++ packages/melt/src/lib/builders/index.ts
@@ -1,5 +1,7 @@
+export * from './Collapsible.svelte';
 export * from "./PinInput.svelte";
 export * from "./Popover.svelte";
+export * from "./Progress.svelte";
 export * from "./RadioGroup.svelte";
 export * from "./Tabs.svelte";
 export * from "./Toggle.svelte";
@@ -7,3 +9,4 @@ export * from "./Slider.svelte";
 export * from "./utils.svelte";
 export * from "./Tooltip.svelte";
 export * from "./Tree.svelte";
+export * from "./Avatar.svelte";