+++ packages/melt/CHANGELOG.md
@@ -1,5 +1,29 @@
 # melt
 
+## 0.6.0
+
+### Minor Changes
+
+- 3a7317e: feat: add avatar
+
+## 0.5.1
+
+### Patch Changes
+
+- 33e8480: Update peer dependency and libs
+
+## 0.5.0
+
+### Minor Changes
+
+- 827e5bb: feat: add collapsible builder
+
+## 0.4.0
+
+### Minor Changes
+
+- 7b165bc: add progress builder
+
 ## 0.3.0
 
 ### Minor Changes